Lock / unlock may be called multiple times

+++ R/UTILS-lockfile.R
@@ -1,17 +1,33 @@
+lockfile_counter = new.env()
+
 #' Create a directory to act as a lockfile, error if the file already exists
 #'
 #' @param path Path to the lockfile
 #' @param timeout How many times to attempt to obtain the lock
+#' @param unique The lock may only be obtained once during this session
 #'
 #' @export
 #' @rdname lockfile
-make_lockfile <- function(path,timeout=10) {
+make_lockfile <- function(path,timeout=10, unique=FALSE) {
   if(is.na(path)) stop("Unable to obtain lock: Path is invalid")
-
+  pname=make.names(normalizePath(path,mustWork = FALSE))
+  counter=lockfile_counter[[pname]]
+  if(length(counter) && counter>0) {
+    log.debug('lock counter: %s=%s',pname,counter)
+    lockfile_counter[[pname]]=counter+1
+    if(!unique) return(invisible(TRUE))
+  } else {
+    counter=0
+    lockfile_counter[[pname]]=0
+  }
+  log.debug('lock counter: %s=%s',pname,counter)
   warning_shown=FALSE
   for(i in 1:timeout) {
     res=dir.create(path,showWarnings=FALSE)
-    if(res==TRUE) return(invisible(TRUE))
+    if(res==TRUE) {
+      lockfile_counter[[pname]]=counter+1
+      return(invisible(TRUE))
+    }
     if(! warning_shown) {
       warning("Unable to obtain lock, retrying for ",timeout, " seconds.",immediate. = TRUE)
       warning_shown=TRUE
@@ -25,6 +41,16 @@ make_lockfile <- function(path,timeout=10) {
 #' @rdname lockfile
 remove_lockfile <- function(path) {
   if(is.na(path)) stop("Unable to release lock: Path is invalid")
+  pname=make.names(normalizePath(path,mustWork = FALSE))
+  counter=lockfile_counter[[pname]]
+  log.debug('Using lock counter: %s',pname)
+  if(!length(counter)) stop('Unable to release lock: lock was never obtained')
+  log.debug('lock counter: %s=%s',pname,counter)
+  if(counter>1) {
+    lockfile_counter[[pname]]=counter-1
+    return(invisible(TRUE))
+  }
+  lockfile_counter[[pname]]=0
   unlink(sprintf(path),recursive=TRUE)
   invisible(TRUE)
 }
